[BUGFIX] Reset RteHtmlParser properly when configuration changes
Before this fix, only the first RTE processing configuration would be
taken into account when deciding which tags to keep in an HTML snippet.

RteHtmlParser is a DI service and as such is treated as a singleton.
Using multiple RTE processing configurations with potentially different
sets of allowed tags should work as expected.

Resolves: #101190
Releases: main, 12.4, 11.5
Change-Id: Ib0227872b542f53b30a16494aa2f979b8f601e07